+++ Sources/Collections/SkeletonCollectionDataSource.swift
@@ -38,6 +38,7 @@ extension SkeletonCollectionDataSource: UITableViewDataSource {
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cellIdentifier = originalTableViewDataSource?.collectionSkeletonView(tableView, cellIdentifierForRowAt: indexPath) ?? ""
         let cell = tableView.dequeueReusableCell(withIdentifier: cellIdentifier, for: indexPath)
+        skeletonCellIfContainerSkeletonIsActive(container: tableView, cell: cell)
         return cell
     }
 }
@@ -56,6 +57,7 @@ extension SkeletonCollectionDataSource: UICollectionViewDataSource {
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let cellIdentifier = originalCollectionViewDataSource?.collectionSkeletonView(collectionView, cellIdentifierForItemAt: indexPath) ?? ""
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellIdentifier, for: indexPath)
+        skeletonCellIfContainerSkeletonIsActive(container: collectionView, cell: cell)
         return cell
     }
     
@@ -71,4 +73,15 @@ extension SkeletonCollectionDataSource: UICollectionViewDataSource {
         return UICollectionReusableView()
     }
     
+}
+
+extension SkeletonCollectionDataSource {
+    private func skeletonCellIfContainerSkeletonIsActive(container: UIView, cell: UIView) {
+        guard container.isSkeletonActive,
+              let skeletonConfig = container.currentSkeletonConfig else {
+            return
+        }
+
+        cell.showSkeleton(skeletonConfig: skeletonConfig)
+    }
 }