+++ Sources/Collections/UIView+CollectionSkeleton.swift
@@ -15,6 +15,11 @@ extension UIView {
         collection.disableUserInteraction()
     }
     
+    func updateDummyDataSourceIfNeeded() {
+        guard let collection = self as? CollectionSkeleton else { return }
+        collection.updateDummyDataSource()
+    }
+    
     func removeDummyDataSourceIfNeeded(reloadAfter reload: Bool = true) {
         guard let collection = self as? CollectionSkeleton else { return }
         collection.removeDummyDataSource(reloadAfter: reload)