+++ Sources/Extensions/CALayer+Extensions.swift
@@ -46,7 +46,7 @@ extension CALayer {
             .setCornerRadius(multilineCornerRadius)
 
         (0..<numberOfSublayers).forEach { index in
-            var width = bounds.width
+            var width = getLineWidth(index: index, numberOfSublayers: numberOfSublayers, lastLineFillPercent: lastLineFillPercent)
             if index == numberOfSublayers - 1 && numberOfSublayers != 1 {
                 width = width * CGFloat(lastLineFillPercent) / 100;
             }
@@ -60,6 +60,29 @@ extension CALayer {
         }
     }
     
+    func updateMultilinesLayers(lastLineFillPercent: Int) {
+        let currentSkeletonSublayers = skeletonSublayers
+        let numberOfSublayers = currentSkeletonSublayers.count
+        for (index, layer) in currentSkeletonSublayers.enumerated() {
+            let width = getLineWidth(index: index, numberOfSublayers: numberOfSublayers, lastLineFillPercent: lastLineFillPercent)
+            layer.updateLayerFrame(for: index, width: width)
+        }
+    }
+
+    private func getLineWidth(index: Int, numberOfSublayers: Int, lastLineFillPercent: Int) -> CGFloat {
+        var width = bounds.width
+        if index == numberOfSublayers - 1 && numberOfSublayers != 1 {
+            width = width * CGFloat(lastLineFillPercent) / 100;
+        }
+
+        return width
+    }
+
+    func updateLayerFrame(for index: Int, width: CGFloat) {
+        let spaceRequiredForEachLine = SkeletonAppearance.default.multilineHeight + SkeletonAppearance.default.multilineSpacing
+        frame = CGRect(x: 0.0, y: CGFloat(index) * spaceRequiredForEachLine, width: width, height: SkeletonAppearance.default.multilineHeight)
+    }
+
     private func calculateNumLines(maxLines: Int) -> Int {
         let spaceRequitedForEachLine = SkeletonAppearance.default.multilineHeight + SkeletonAppearance.default.multilineSpacing
         var numberOfSublayers = Int(round(CGFloat(bounds.height)/CGFloat(spaceRequitedForEachLine)))