+++ Sources/SkeletonConfig.swift
@@ -1,0 +1,31 @@
+// Copyright © 2018 SkeletonView. All rights reserved.
+
+import UIKit
+
+/// Used to store all config needed to activate the skeleton layer.
+struct SkeletonConfig {
+    /// Type of skeleton layer
+    let type: SkeletonType
+    /// Colors used in skeleton layer
+    let colors: [UIColor]
+    /// If type is gradient, which gradient direction
+    let gradientDirection: GradientDirection?
+    /// Specify if skeleton is animated or not
+    let animated: Bool
+    /// Used to execute a custom animation
+    let animation: SkeletonLayerAnimation?
+
+    init(
+            type: SkeletonType,
+            colors: [UIColor],
+            gradientDirection: GradientDirection? = nil,
+            animated: Bool = false,
+            animation: SkeletonLayerAnimation? = nil
+    ) {
+        self.type = type
+        self.colors = colors
+        self.gradientDirection = gradientDirection
+        self.animated = animated
+        self.animation = animation
+    }
+}