+++ Sources/SkeletonLayer.swift
@@ -55,6 +55,18 @@ struct SkeletonLayer {
         self.maskLayer.tint(withColors: colors)
     }
     
+    func update(usingColors colors: [UIColor]) {
+        layoutIfNeeded()
+        self.maskLayer.tint(withColors: colors)
+    }
+
+    func layoutIfNeeded() {
+        if let bounds = self.holder?.maxBoundsEstimated {
+            self.maskLayer.bounds = bounds
+        }
+        updateMultilinesIfNeeded()
+    }
+    
     func removeLayer() {
         maskLayer.removeFromSuperlayer()
     }
@@ -62,6 +74,11 @@ struct SkeletonLayer {
     func addMultilinesIfNeeded() {
         guard let multiLineView = holder as? ContainsMultilineText else { return }
         maskLayer.addMultilinesLayers(lines: multiLineView.numLines, type: type, lastLineFillPercent: multiLineView.lastLineFillingPercent, multilineCornerRadius: multiLineView.multilineCornerRadius)
+    }
+
+    func updateMultilinesIfNeeded() {
+        guard let multiLineView = holder as? ContainsMultilineText else { return }
+        maskLayer.updateMultilinesLayers(lastLineFillPercent: multiLineView.lastLineFillingPercent)
     }
 }